This is my Flutter code for Firebase Push Notifications.
In this code the Notifications are displayed when app is in Foreground, background or killed.

When the notification is clicked so certain action is performed, i.e., redirecting user to next screen while passing the data/payload coming from Firebase Cloud Messaging (FCM), with it as well.

The user can also send notifications from one device to another using the Token of the recipient device.